@@ -21,60 +22,61 @@
   "dependencies": {
     "@ant-design/icons": "^4.6.2",
     "@apollo/client": "^3.3.20",
-    "@babel/core": "7.12.3",
+    "@babel/core": "^7.16.0",
     "@civic/solana-gateway-react": "^0.4.13",
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
     "@ledgerhq/hw-transport-webusb": "^6.0.2",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
     "@metaplex-foundation/mpl-token-metadata": "^1.2.5",
     "@metaplex/js": "^4.10.1",
-    "@mui/core": "^5.0.0-alpha.54",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/lab": "^5.0.0-alpha.70",
-    "@mui/material": "^5.4.3",
-    "@pmmmwh/react-refresh-webpack-plugin": "0.4.3",
-    "@project-serum/serum": "^0.13.41",
+    "@mui/icons-material": "^5.6.2",
+    "@mui/lab": "^5.0.0-alpha.80",
+    "@mui/material": "^5.6.4",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
+    "@project-serum/anchor": "^0.24.2",
+    "@project-serum/serum": "^0.13.65",
     "@project-serum/sol-wallet-adapter": "^0.2.5",
     "@sentry/react": "^6.13.3",
     "@sentry/tracing": "^6.13.3",
     "@solana/spl-token": "^0.1.8",
     "@solana/spl-token-registry": "^0.2.168",
     "@solana/spl-token-swap": "0.1.0",
-    "@solana/wallet-adapter-base": "^0.5.2",
-    "@solana/wallet-adapter-material-ui": "^0.13.0",
-    "@solana/wallet-adapter-react": "^0.9.2",
-    "@solana/wallet-adapter-react-ui": "^0.1.2",
-    "@solana/wallet-adapter-wallets": "^0.11.3",
+    "@solana/wallet-adapter-base": "^0.9.5",
+    "@solana/wallet-adapter-material-ui": "^0.16.5",
+    "@solana/wallet-adapter-react": "^0.15.4",
+    "@solana/wallet-adapter-react-ui": "^0.9.6",
+    "@solana/wallet-adapter-wallets": "^0.16.0",
     "@solana/web3.js": "^1.32.0",
-    "@svgr/webpack": "5.5.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.1.9",
+    "@svgr/webpack": "^5.5.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^13.5.0",
     "@typescript-eslint/eslint-plugin": "^4.5.0",
     "@typescript-eslint/parser": "^4.5.0",
     "axios": "0.21.1",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "8.1.0",
-    "babel-plugin-named-asset-import": "^0.3.7",
-    "babel-preset-react-app": "^10.0.0",
+    "babel-jest": "^27.4.2",
+    "babel-loader": "^8.2.3",
+    "babel-plugin-named-asset-import": "^0.3.8",
+    "babel-preset-react-app": "^10.0.1",
     "bfj": "^7.0.2",
     "bigint-buffer": "^1.1.5",
     "bignumber.js": "^9.0.1",
     "bn.js": "^5.2.0",
+    "browserslist": "^4.18.1",
     "bs58": "^4.0.1",
     "buffer-layout": "^1.2.2",
-    "camelcase": "^6.1.0",
-    "case-sensitive-paths-webpack-plugin": "2.3.0",
-    "css-loader": "4.3.0",
+    "camelcase": "^6.2.1",
+    "case-sensitive-paths-webpack-plugin": "^2.4.0",
+    "crypto-browserify": "^3.12.0",
+    "css-loader": "^6.5.1",
+    "css-minimizer-webpack-plugin": "^3.2.0",
     "cuid": "^2.1.8",
-    "dotenv": "8.2.0",
-    "dotenv-expand": "5.1.0",
+    "dotenv": "^10.0.0",
+    "dotenv-expand": "^5.1.0",
     "emoji-mart": "^3.0.1",
-    "eslint": "^7.11.0",
-    "eslint-config-react-app": "^6.0.0",
+    "eslint": "^8.3.0",
+    "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-flowtype": "^5.2.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jest": "^24.1.0",
@@ -82,69 +84,72 @@
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-testing-library": "^3.9.2",
-    "eslint-webpack-plugin": "^2.5.2",
+    "eslint-webpack-plugin": "^3.1.1",
     "eventemitter3": "^4.0.7",
-    "file-loader": "6.1.1",
-    "fs-extra": "^9.0.1",
+    "file-loader": "^6.2.0",
+    "fs-extra": "^10.0.0",
     "graphql": "^15.5.1",
-    "html-webpack-plugin": "4.5.0",
+    "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.0",
-    "identity-obj-proxy": "3.0.0",
+    "identity-obj-proxy": "^3.0.0",
     "jazzicon": "^1.5.0",
-    "jest": "26.6.0",
+    "jest": "^27.4.3",
     "jest-circus": "26.6.0",
-    "jest-resolve": "26.6.0",
-    "jest-watch-typeahead": "0.6.1",
+    "jest-resolve": "^27.4.2",
+    "jest-watch-typeahead": "^1.0.0",
     "linkify-plugin-hashtag": "^3.0.3",
     "linkify-plugin-mention": "^3.0.3",
     "linkify-react": "^3.0.3",
     "linkifyjs": "^3.0.5",
     "lodash": "^4.17.21",
-    "mini-css-extract-plugin": "0.11.3",
+    "mini-css-extract-plugin": "^2.4.5",
     "moment": "2.29.1",
     "mz": "^2.7.0",
-    "notistack": "^1.0.9",
     "optimize-css-assets-webpack-plugin": "5.0.4",
     "pnp-webpack-plugin": "1.6.4",
-    "postcss-flexbugs-fixes": "4.2.1",
-    "postcss-loader": "3.0.0",
-    "postcss-normalize": "8.0.1",
-    "postcss-preset-env": "6.7.0",
+    "postcss": "^8.4.4",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-loader": "^6.2.1",
+    "postcss-normalize": "^10.0.1",
+    "postcss-preset-env": "^7.0.1",
     "postcss-safe-parser": "5.0.2",
     "pretty-quick": "^3.1.1",
-    "prompts": "2.4.0",
-    "react": "^17.0.2",
+    "prompts": "^2.4.2",
+    "react": "^18.1.0",
     "react-apollo": "3.1.5",
-    "react-app-polyfill": "^2.0.0",
+    "react-app-polyfill": "^3.0.0",
     "react-countdown": "^2.3.2",
-    "react-dev-utils": "^11.0.3",
-    "react-dom": "^17.0.2",
+    "react-dev-utils": "^12.0.1",
+    "react-dom": "^18.1.0",
     "react-helmet": "^6.1.0",
     "react-infinite-scroll-component": "^6.1.0",
     "react-onesignal": "^2.0.2",
     "react-parallax-tilt": "^1.6.1",
-    "react-refresh": "^0.8.3",
+    "react-refresh": "^0.11.0",
     "react-router-dom": "^6.0.2",
     "react-router-hash-link": "^2.4.3",
     "react-textarea-autosize": "8.3.3",
     "reactjs-popup": "^2.0.5",
-    "recoil": "^0.6.1",
-    "resolve": "1.18.1",
-    "resolve-url-loader": "^3.1.2",
-    "sass-loader": "^10.0.5",
-    "semver": "7.3.2",
+    "recoil": "^0.7.3-alpha.1",
+    "resolve": "^1.20.0",
+    "resolve-url-loader": "^4.0.0",
+    "sass-loader": "^12.3.0",
+    "semver": "^7.3.5",
     "simple-react-lightbox": "^3.6.9-0",
-    "style-loader": "1.3.0",
-    "terser-webpack-plugin": "4.2.3",
+    "source-map-loader": "^3.0.0",
+    "style-loader": "^3.3.1",
+    "tailwindcss": "^3.0.2",
+    "terser-webpack-plugin": "^5.2.5",
     "ts-node": "^10.1.0",
     "ts-pnp": "1.2.0",
     "typescript": "^4.3.5",
     "unfurl.js": "^5.6.4",
     "url-loader": "4.1.1",
-    "webpack": "4.44.2",
-    "webpack-dev-server": "3.11.1",
-    "webpack-manifest-plugin": "2.2.0",
-    "workbox-webpack-plugin": "5.1.4"
+    "web-vitals": "^2.1.4",
+    "webpack": "^5.72.0",
+    "webpack-dev-server": "^4.6.0",
+    "webpack-manifest-plugin": "^4.0.2",
+    "workbox-webpack-plugin": "^6.4.1"
   },
   "devDependencies": {
     "@babel/plugin-transform-modules-commonjs": "^7.17.7",
