@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useState, useCallback, useContext } from "react";
 import * as anchor from "@project-serum/anchor";
-import { Container, Snackbar } from "@material-ui/core";
+import { Container, Snackbar } from "@mui/material";
 import CheckIcon from "@mui/icons-material/Check";
 import AccountBalanceWalletIcon from '@mui/icons-material/AccountBalanceWallet';
 import CableIcon from '@mui/icons-material/Cable';
@@ -133,14 +133,12 @@ export const OGMint = () => {
       document.getElementById("#identity")?.click();
 
       if (wallet.connected && candyMachine?.program && wallet.publicKey) {
-        const mintTxId = (
-          await mintOneToken(candyMachine, wallet.publicKey)
-        )[0];
+        const mintResult = await mintOneToken(candyMachine, wallet.publicKey);
 
         let status: any = { err: true };
-        if (mintTxId) {
+        if (mintResult) {
           status = await awaitTransactionSignatureConfirmation(
-            mintTxId,
+            mintResult?.mintTxId,
             txTimeoutInMilliseconds,
             connection,
             true
@@ -231,6 +229,7 @@ export const OGMint = () => {
                     candyMachine?.state.gatekeeper &&
                     wallet.publicKey &&
                     wallet.signTransaction ? (
+                      //@ts-ignore
                       <GatewayProvider
                         wallet={{
                           publicKey:
