@@ -16,7 +16,7 @@ export const UserContext = React.createContext<UserContextType>({
   setUser: () => console.warn("no setUser function provided"),
 });
 
-export const UserContextProvider: React.FC = ({ children }) => {
+export const UserContextProvider: React.FC<{children: JSX.Element}> = ({ children }) => {
   const [user, _setUser] = useState<User | null>(null);
 
   useEffect(() => {
