@@ -1,16 +1,21 @@
 import * as anchor from "@project-serum/anchor";
-
-import { MintLayout, TOKEN_PROGRAM_ID, Token } from "@solana/spl-token";
-import { SystemProgram } from "@solana/web3.js";
-import { sendTransactions } from "../contexts/connection";
-
 import {
   CIVIC,
+  SPL_ASSOCIATED_TOKEN_ACCOUNT_PROGRAM_ID,
   getAtaForMint,
   getNetworkExpire,
   getNetworkToken,
-  SPL_ASSOCIATED_TOKEN_ACCOUNT_PROGRAM_ID,
 } from "./utils";
+import { MintLayout, TOKEN_PROGRAM_ID, Token } from "@solana/spl-token";
+import {
+  SYSVAR_SLOT_HASHES_PUBKEY,
+  SystemProgram,
+  Transaction,
+} from "@solana/web3.js";
+import {
+  sendTransactions,
+  SequenceType,
+} from "../contexts/connection";
 
 export const CANDY_MACHINE_PROGRAM = new anchor.web3.PublicKey(
   "cndy3Z4yapfJBmL3ShUp5exZKqR3z33thTzeNMm2gRZ"
@@ -21,21 +26,26 @@ const TOKEN_METADATA_PROGRAM_ID = new anchor.web3.PublicKey(
 );
 
 interface CandyMachineState {
+  authority: anchor.web3.PublicKey;
   itemsAvailable: number;
   itemsRedeemed: number;
   itemsRemaining: number;
   treasury: anchor.web3.PublicKey;
-  tokenMint: anchor.web3.PublicKey;
+  tokenMint: null | anchor.web3.PublicKey;
   isSoldOut: boolean;
   isActive: boolean;
   isPresale: boolean;
+  isWhitelistOnly: boolean;
   goLiveDate: anchor.BN;
   price: anchor.BN;
   gatekeeper: null | {
     expireOnUse: boolean;
     gatekeeperNetwork: anchor.web3.PublicKey;
   };
-  endSettings: null | [number, anchor.BN];
+  endSettings: null | {
+    number: anchor.BN;
+    endSettingType: any;
+  };
   whitelistMintSettings: null | {
     mode: any;
     mint: anchor.web3.PublicKey;
@@ -47,6 +57,7 @@ interface CandyMachineState {
     uri: string;
     hash: Uint8Array;
   };
+  retainAuthority: boolean;
 }
 
 export interface CandyMachineAccount {
@@ -155,43 +166,31 @@ export const getCandyMachineState = async (
   candyMachineId: anchor.web3.PublicKey,
   connection: anchor.web3.Connection
 ): Promise<CandyMachineAccount> => {
-  const provider = new anchor.Provider(connection, anchorWallet, {
-    preflightCommitment: "recent",
+  const provider = new anchor.AnchorProvider(connection, anchorWallet, {
+    preflightCommitment: "processed",
   });
 
   const idl = await anchor.Program.fetchIdl(CANDY_MACHINE_PROGRAM, provider);
 
-  const program = new anchor.Program(idl, CANDY_MACHINE_PROGRAM, provider);
+  const program = new anchor.Program(idl!, CANDY_MACHINE_PROGRAM, provider);
 
   const state: any = await program.account.candyMachine.fetch(candyMachineId);
   const itemsAvailable = state.data.itemsAvailable.toNumber();
   const itemsRedeemed = state.itemsRedeemed.toNumber();
   const itemsRemaining = itemsAvailable - itemsRedeemed;
 
-  const presale =
-    state.data.whitelistMintSettings &&
-    state.data.whitelistMintSettings.presale &&
-    (!state.data.goLiveDate ||
-      state.data.goLiveDate.toNumber() > new Date().getTime() / 1000);
-
   return {
     id: candyMachineId,
     program,
     state: {
+      authority: state.authority,
       itemsAvailable,
       itemsRedeemed,
       itemsRemaining,
       isSoldOut: itemsRemaining === 0,
-      isActive:
-        (presale ||
-          state.data.goLiveDate.toNumber() < new Date().getTime() / 1000) &&
-        (state.data.endSettings
-          ? state.data.endSettings.endSettingType.date
-            ? state.data.endSettings.number.toNumber() >
-              new Date().getTime() / 1000
-            : itemsRedeemed < state.data.endSettings.number.toNumber()
-          : true),
-      isPresale: presale,
+      isActive: false,
+      isPresale: false,
+      isWhitelistOnly: false,
       goLiveDate: state.data.goLiveDate,
       treasury: state.wallet,
       tokenMint: state.tokenMint,
@@ -200,6 +199,7 @@ export const getCandyMachineState = async (
       whitelistMintSettings: state.data.whitelistMintSettings,
       hiddenSettings: state.data.hiddenSettings,
       price: state.data.price,
+      retainAuthority: state.data.retainAuthority,
     },
   };
 };
@@ -244,24 +244,54 @@ export const getCandyMachineCreator = async (
   );
 };
 
-export const mintOneToken = async (
+export const getCollectionPDA = async (
+  candyMachineAddress: anchor.web3.PublicKey
+): Promise<[anchor.web3.PublicKey, number]> => {
+  return await anchor.web3.PublicKey.findProgramAddress(
+    [Buffer.from("collection"), candyMachineAddress.toBuffer()],
+    CANDY_MACHINE_PROGRAM
+  );
+};
+
+export interface CollectionData {
+  mint: anchor.web3.PublicKey;
+  candyMachine: anchor.web3.PublicKey;
+}
+
+export const getCollectionAuthorityRecordPDA = async (
+  mint: anchor.web3.PublicKey,
+  newAuthority: anchor.web3.PublicKey
+): Promise<anchor.web3.PublicKey> => {
+  return (
+    await anchor.web3.PublicKey.findProgramAddress(
+      [
+        Buffer.from("metadata"),
+        TOKEN_METADATA_PROGRAM_ID.toBuffer(),
+        mint.toBuffer(),
+        Buffer.from("collection_authority"),
+        newAuthority.toBuffer(),
+      ],
+      TOKEN_METADATA_PROGRAM_ID
+    )
+  )[0];
+};
+
+export type SetupState = {
+  mint: anchor.web3.Keypair;
+  userTokenAccount: anchor.web3.PublicKey;
+  transaction: string;
+};
+
+export const createAccountsForMint = async (
   candyMachine: CandyMachineAccount,
   payer: anchor.web3.PublicKey
-): Promise<(string | undefined)[]> => {
+): Promise<SetupState> => {
   const mint = anchor.web3.Keypair.generate();
-
   const userTokenAccountAddress = (
     await getAtaForMint(mint.publicKey, payer)
   )[0];
 
-  const userPayingAccountAddress = candyMachine.state.tokenMint
-    ? (await getAtaForMint(candyMachine.state.tokenMint, payer))[0]
-    : payer;
-
-  const candyMachineAddress = candyMachine.id;
-  const remainingAccounts = [];
   const signers: anchor.web3.Keypair[] = [mint];
-  const cleanupInstructions = [];
   const instructions = [
     anchor.web3.SystemProgram.createAccount({
       fromPubkey: payer,
@@ -296,6 +326,91 @@ export const mintOneToken = async (
     ),
   ];
 
+  return {
+    mint: mint,
+    userTokenAccount: userTokenAccountAddress,
+    transaction: (
+      await sendTransactions(
+        candyMachine.program.provider.connection,
+        payer,
+        [instructions],
+        [signers],
+        SequenceType.StopOnFailure,
+        "singleGossip",
+        () => {},
+        () => false,
+        undefined
+      )
+    ).txs[0].txid,
+  };
+};
+
+type MintResult = {
+  mintTxId: string;
+  metadataKey: anchor.web3.PublicKey;
+};
+
+export const mintOneToken = async (
+  candyMachine: CandyMachineAccount,
+  payer: anchor.web3.PublicKey,
+  beforeTransactions: Transaction[] = [],
+  afterTransactions: Transaction[] = [],
+  setupState?: SetupState
+): Promise<MintResult | null> => {
+  const mint = setupState?.mint ?? anchor.web3.Keypair.generate();
+  const userTokenAccountAddress = (
+    await getAtaForMint(mint.publicKey, payer)
+  )[0];
+
+  const userPayingAccountAddress = candyMachine.state.tokenMint
+    ? (await getAtaForMint(candyMachine.state.tokenMint, payer))[0]
+    : payer;
+
+  const candyMachineAddress = candyMachine.id;
+  const remainingAccounts = [];
+  const cleanupInstructions = [];
+  const instructions = [];
+  const signers: anchor.web3.Keypair[] = [];
+  console.log("SetupState: ", setupState);
+  if (!setupState) {
+    signers.push(mint);
+    instructions.push(
+      ...[
+        anchor.web3.SystemProgram.createAccount({
+          fromPubkey: payer,
+          newAccountPubkey: mint.publicKey,
+          space: MintLayout.span,
+          lamports:
+            await candyMachine.program.provider.connection.getMinimumBalanceForRentExemption(
+              MintLayout.span
+            ),
+          programId: TOKEN_PROGRAM_ID,
+        }),
+        Token.createInitMintInstruction(
+          TOKEN_PROGRAM_ID,
+          mint.publicKey,
+          0,
+          payer,
+          payer
+        ),
+        createAssociatedTokenAccountInstruction(
+          userTokenAccountAddress,
+          payer,
+          payer,
+          mint.publicKey
+        ),
+        Token.createMintToInstruction(
+          TOKEN_PROGRAM_ID,
+          mint.publicKey,
+          userTokenAccountAddress,
+          payer,
+          [],
+          1
+        ),
+      ]
+    );
+  }
+
   if (candyMachine.state.gatekeeper) {
     remainingAccounts.push({
       pubkey: (
@@ -307,6 +422,7 @@ export const mintOneToken = async (
       isWritable: true,
       isSigner: false,
     });
+
     if (candyMachine.state.gatekeeper.expireOnUse) {
       remainingAccounts.push({
         pubkey: CIVIC,
@@ -418,6 +534,7 @@ export const mintOneToken = async (
     candyMachineAddress
   );
 
+  console.log(remainingAccounts.map((rm) => rm.pubkey.toBase58()));
   instructions.push(
     await candyMachine.program.instruction.mintNft(creatorBump, {
       accounts: {
@@ -435,7 +552,7 @@ export const mintOneToken = async (
         systemProgram: SystemProgram.programId,
         rent: anchor.web3.SYSVAR_RENT_PUBKEY,
         clock: anchor.web3.SYSVAR_CLOCK_PUBKEY,
-        recentBlockhashes: anchor.web3.SYSVAR_RECENT_BLOCKHASHES_PUBKEY,
+        recentBlockhashes: SYSVAR_SLOT_HASHES_PUBKEY,
         instructionSysvarAccount: anchor.web3.SYSVAR_INSTRUCTIONS_PUBKEY,
       },
       remainingAccounts:
@@ -443,20 +560,83 @@ export const mintOneToken = async (
     })
   );
 
+  const [collectionPDA] = await getCollectionPDA(candyMachineAddress);
+  const collectionPDAAccount =
+    await candyMachine.program.provider.connection.getAccountInfo(
+      collectionPDA
+    );
+
+  if (collectionPDAAccount && candyMachine.state.retainAuthority) {
+    try {
+      const collectionData =
+        (await candyMachine.program.account.collectionPda.fetch(
+          collectionPDA
+        )) as unknown as CollectionData;
+      console.log(collectionData);
+      const collectionMint = collectionData.mint;
+      const collectionAuthorityRecord = await getCollectionAuthorityRecordPDA(
+        collectionMint,
+        collectionPDA
+      );
+      console.log(collectionMint);
+      if (collectionMint) {
+        const collectionMetadata = await getMetadata(collectionMint);
+        const collectionMasterEdition = await getMasterEdition(collectionMint);
+        console.log("Collection PDA: ", collectionPDA.toBase58());
+        console.log("Authority: ", candyMachine.state.authority.toBase58());
+        instructions.push(
+          await candyMachine.program.instruction.setCollectionDuringMint({
+            accounts: {
+              candyMachine: candyMachineAddress,
+              metadata: metadataAddress,
+              payer: payer,
+              collectionPda: collectionPDA,
+              tokenMetadataProgram: TOKEN_METADATA_PROGRAM_ID,
+              instructions: anchor.web3.SYSVAR_INSTRUCTIONS_PUBKEY,
+              collectionMint,
+              collectionMetadata,
+              collectionMasterEdition,
+              authority: candyMachine.state.authority,
+              collectionAuthorityRecord,
+            },
+          })
+        );
+      }
+    } catch (error) {
+      console.error(error);
+    }
+  }
+
+  const instructionsMatrix = [instructions, cleanupInstructions];
+  const signersMatrix = [signers, []];
+
   try {
-    return (
+    const txns = (
       await sendTransactions(
         candyMachine.program.provider.connection,
-        candyMachine.program.provider.wallet,
-        [instructions, cleanupInstructions],
-        [signers, []]
+        payer,
+        instructionsMatrix,
+        signersMatrix,
+        SequenceType.StopOnFailure,
+        "singleGossip",
+        () => {},
+        () => false,
+        undefined
       )
-    ).txs.map((t) => t.txid);
+    ).txs.map((t: any) => t.txid);
+    const mintTxn = txns[0];
+    return {
+      mintTxId: mintTxn,
+      metadataKey: metadataAddress,
+    };
   } catch (e) {
     console.log(e);
   }
+  return null;
+};
 
-  return [];
+export const shortenAddress = (address: string, chars = 4): string => {
+  return `${address.slice(0, chars)}...${address.slice(-chars)}`;
 };
 
 const sleep = (ms: number): Promise<void> => {
