@@ -4,7 +4,7 @@ import AutoGraphIcon from "@mui/icons-material/AutoGraph";
 import GroupIcon from "@mui/icons-material/Group";
 import Language from "@mui/icons-material/Language";
 import NotificationsIcon from "@mui/icons-material/Notifications";
-import StyleIcon from "@mui/icons-material/Style";
+// import StyleIcon from "@mui/icons-material/Style";
 import TagIcon from "@mui/icons-material/Tag";
 import { Avatar, Badge, Stack, Typography } from "@mui/material";
 import { Box } from "@mui/system";
@@ -17,7 +17,7 @@ import { styled } from "@mui/material/styles";
 import { useChannelsQuery } from "../generated/graphql";
 
 interface Props {
-  newMentionsCount: Number | undefined;
+  newMentionsCount: number | undefined;
   user: StoreUser | null;
 }
 
